Extract Game, Player and Card to leanpoker package
