+++ leanpoker/card.go
@@ -1,0 +1,9 @@
+package leanpocker
+
+type Card struct {
+	// Rank of the card. Possible values are numbers 2-10 and J,Q,K,A
+	Rank string `json:"rank"`
+
+	// Suit of the card. Possible values are: clubs,spades,hearts,diamonds
+	Suit string `json:"suit"`
+}
